README: remove call to action about other languages
* After a lot of thought and exploring many languages, I
  have yet to find another language where the errnoname
  approach (generating library code from a list of errno
  names) would works in a useful way.

* In C, `errnoname` can just be ignorant of what value
  each `errno` name maps to, or if it is even defined.
  The C preprocessor and the system-provided `errno.h`
  header figure out the rest. The piece that's missing
  in C is knowing all the names that are possible, and
  exposing them as strings to other code. The current
  `errnoname` approach provides exactly that.

* All other languages I have looked at have a different
  problem, one which can't be solved by knowing all the
  possible names. In those languages, `errnoname` would
  need to know the numbers, because the system does not
  provide the number-to-name mapping to that language
  in that language like it does for C.

So to provide an implementation of `errnoname` in those
languages, the approach would need to be very different.
This project is in a good position to do that only to
the extent that it already scrapes every public `errno`
definition that I've been able to find. But the goal of
precisely what information is needed is not a strict
subset in either direction: `errnoname` on C benefits
from `errno` names even if it does not know the value,
which is useless to `errnoname` on every other language.
Conversely, in every other language we need to know the
values for each name for each system, which is useless
detail to `errnoname` on C.

In short, it might still be worth doing, and maybe even
in this project or by reusing pieces of it, but this is
no longer so unambiguously worth doing in this repo to
keep the "let's expand this to cover every language
that needs it!" sentiment on display.