Adding errno names starting with "E" from HaikuOS
Why not also include HaikuOS errno names that start with
"B_" or "HAIKU_"? Because I think the "right way" to solve
that problem is a separate copy of `errnoname` called
`errnoname_haiku` which just covers the "native" errno names
from HaikuOS, and excludes the "E" names. Why? Because there
are good reasons for wanting error names in either form,
those good reasons are only known by code which calls into
`errnoname`, and my intuition is that most of the time the
caller will know which one they want. And we can build
"smarter" stuff on top of those functions, like:

char const * errnoname_prefer_haiku(int errno_)
{
    char const * result = errnoname_haiku(errno_);
    if(!result)
    {
        result = errnoname(errno_);
    }
    return result;
}

The current pre-build script `generate-c.sh` can take any
list of errno names on standard input, so it would be
trivial to feed a list of the "B_" and/or "HAIKU_" names
into it, and then you just need to rename `errnoname` in
`errnoname.c` file before building it.