gather-errno-names.sh: prefix gathering errors by source
This is relatively minor and arguably not worth it (which is why I
did not bother implementing it up until this point) but it really
helps when there is an error message to know which of the sources'
functions that error came from.

The performance-mindful part of me also doesn't love adding an
extra `| sed` invocation for every download function, but this
seems worth the trade, especially because there are not many
use cases for anyone but maintainers of errnoname to run this
script, even less use cases for running it on hardware that is
so severely constrained that this is even noticeable overhead
(let alone a problem), and if it ultimately has to be done, it
should be trivial to just delete `execute_with_prefixed_errors`
from the code.