gather-errno-names.sh: adjusting order to match improvement
I have an informal ordering for the function definitions inside
this file: the downloads of `errno` lists which are worse (such
as requiring parsing a web page just to find the latest version
of the documentation) go below the nicer ones.

z/OS has just elevated itself to basically top-tier (or as top
tier as it can get without its errno headers being open source).