z/OS errno updates
It seems that with the latest z/OS errno updates, lots of errors
have been removed and some have been added. Possibly some were
renamed too. The new documentation says that it excludes errno
codes which are internal to z/OS and subject to change, and the
IBM documentation has made a huge change for the better overall,
so I am just trusting that all of these `errno` name changes are
in our best interests are consumers of errors.