Make HP-UX errno gather more comprehensive
I was and remain conflicted about this change.
The ideal is scraping a URL that GETS UPDATED
when HP releases a new version of HP-UX. The
manpage mirror seemed most promising for that,
but it still has manpages from HP-UX 11i v2,
and is less complete than the list gathered
from the links being added in this commit.

Without a stays-up-to-date link, a hard-coded
list of all (unique to) HP-UX errno is better
in many ways (probably with a comment listing
these links from which I got those errno).

But running code is more "true" (verifiable,
reproducible) than comments, and a download
from links that do not stay up to date is a
better reminder and motivator towards the
ideal of a link that does stay up to date.